refactor: move page components to views catalog
