+++ assets/js/app.js
@@ -17,7 +17,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import '../css/app.css';
-import Home from './components/Home';
+import Home from './views/Home';
 
 ReactDOM.render(<Router><Home /></Router>, document.getElementById('root'));
