+++ assets/js/views/SetupCheck.js
